refactor(api/applications): refactors documents endpoints to own folder (BOAS-536)
