+++ apps/api/src/server/applications/application/application.validators.js
@@ -6,8 +6,6 @@ import {
 	validationErrorHandlerMissing
 } from '../../middleware/error-handler.js';
 import * as caseRepository from '../../repositories/case.repository.js';
-import * as documentRepository from '../../repositories/document.repository.js';
-import * as folderRepository from '../../repositories/folder.repository.js';
 import * as regionRepository from '../../repositories/region.repository.js';
 import * as serviceCustomerRepository from '../../repositories/service-customer.repository.js';
 import * as subSectorRepository from '../../repositories/sub-sector.repository.js';
@@ -77,26 +75,6 @@ const validateExistingMapZoomLevel = async (value) => {
 
 	if (mapZoomLevel == null) {
 		throw new Error('Unknown map zoom level');
-	}
-};
-
-/**
- * @param {string} value
- * @param {import('express-validator').Meta} meta
- */
-const validateDocumentGUIDBelongsToCase = async (value, { req }) => {
-	const documentGUID = await documentRepository.getByDocumentGUID(value);
-
-	if (documentGUID === null || typeof documentGUID === 'undefined') {
-		throw new Error('DocumentGUID must exist in database');
-	}
-
-	const getCaseById = await folderRepository.getById(documentGUID?.folderId);
-
-	const caseId = getCaseById?.caseId;
-
-	if (Number.parseInt(req.params?.caseId, 10) !== caseId) {
-		throw new Error('GUID must belong to correct case');
 	}
 };
 
@@ -213,41 +191,5 @@ export const validateApplicantId = composeMiddleware(
 
 export const validateGetApplicationQuery = composeMiddleware(
 	query('query').optional({ nullable: true }),
-	validationErrorHandler
-);
-
-/**
- * @param {number} value
- * @param {import('express-validator').Meta} meta
- */
-const validateFolderBelongsToCase = async (value, { req }) => {
-	const folder = await folderRepository.getById(value);
-
-	if (folder === null || typeof folder === 'undefined') {
-		throw new Error('Folder must exist in database');
-	}
-
-	if (folder.caseId !== req.params?.id) {
-		throw new Error('Folder does not belong to case');
-	}
-};
-
-export const validateFolderIds = composeMiddleware(
-	body('*.folderId')
-		.toInt()
-		.custom(validateFolderBelongsToCase)
-		.withMessage('Folder must belong to case'),
-	validationErrorHandler
-);
-
-export const validateDocumentGUID = composeMiddleware(
-	param('documentGUID')
-		.custom(validateDocumentGUIDBelongsToCase)
-		.withMessage('GUID must belong to correct case'),
-	validationErrorHandler
-);
-
-export const validateMachineAction = composeMiddleware(
-	body('machineAction').exists().withMessage('Please provide a value for machine action'),
 	validationErrorHandler
 );
