+++ apps/api/src/server/applications/applications.routes.js
@@ -5,24 +5,19 @@ import {
 	createApplication,
 	getApplicationDetails,
 	startCase,
-	updateApplication,
-	updateDocumentStatus
+	updateApplication
 } from './application/application.controller.js';
 import {
 	validateApplicantId,
 	validateApplicationId,
 	validateCreateUpdateApplication,
-	validateDocumentGUID,
-	validateFolderIds,
-	validateGetApplicationQuery,
-	validateMachineAction
+	validateGetApplicationQuery
 } from './application/application.validators.js';
-import { provideDocumentUploadURLs } from './application/documents/document.controller.js';
 import { documentRoutes } from './application/documents/document.routes.js';
-import { validateDocumentsToUploadProvided } from './application/documents/document.validators.js';
 import { fileFoldersRoutes } from './application/file-folders/folders.routes.js';
 import { caseAdminOfficerRoutes } from './case-admin-officer/case-admin-officer.routes.js';
 import { caseOfficerRoutes } from './case-officer/case-officer.routes.js';
+import { documentsRoutes } from './documents/documents.routes.js';
 import { inspectorRoutes } from './inspector/inspector.routes.js';
 import { regionRoutes } from './region/region.routes.js';
 import { caseSearchRoutes } from './search/case-search.routes.js';
@@ -30,6 +25,8 @@ import { sectorRoutes } from './sector/sector.routes.js';
 import { zoomLevelRoutes } from './zoom-level/zoom-level.routes.js';
 
 const router = createRouter();
+
+router.use('/:id/documents', documentsRoutes);
 
 router.use('/case-officer', caseOfficerRoutes);
 
@@ -146,64 +143,6 @@ router.get(
 	validateGetApplicationQuery,
 	trimUnexpectedRequestParameters,
 	asyncHandler(getApplicationDetails)
-);
-
-router.post(
-	'/:id/documents',
-	/*
-        #swagger.tags = ['Applications']
-        #swagger.path = '/applications/{id}/documents'
-        #swagger.description = 'Saves new documents to database and returns location in Blob Storage'
-        #swagger.parameters['id'] = {
-            in: 'path',
-			description: 'Application ID here',
-			required: true,
-			type: 'integer'
-        }
-        #swagger.parameters['body'] = {
-            in: 'body',
-            description: 'Document Details',
-            schema: { $ref: '#/definitions/documentsToSave' }
-        }
-        #swagger.responses[200] = {
-            description: 'Documents that have been saved',
-            schema: { $ref: '#/definitions/documentsAndBlobStorageURLs' }
-        }
-	 */
-	validateApplicationId,
-	validateDocumentsToUploadProvided,
-	validateFolderIds,
-	trimUnexpectedRequestParameters,
-	asyncHandler(provideDocumentUploadURLs)
-);
-
-router.patch(
-	'/:caseId/documents/:documentGUID/status',
-	/*
-        #swagger.tags = ['Applications']
-        #swagger.path =  'applications/{caseId}/documents/{documentGUID}/status'
-        #swagger.description = 'Updates document status from state machine'
-        #swagger.parameters['caseId'] = {
-            in: 'path',
-			description: 'Case ID here',
-			required: true,
-			type: 'integer'
-        }
-        #swagger.parameters['documentGUID'] = {
-            in: 'path',
-            description: 'Document GUID',
-			required: true,
-			type: 'string'
-        }
-        #swagger.responses[200] = {
-            description: 'Document status updated',
-            schema: { caseId: 1, guid: 'DB0110203', status: 'awaiting_virus_check'}
-        }
-	 */
-	validateDocumentGUID,
-	validateMachineAction,
-	trimUnexpectedRequestParameters,
-	asyncHandler(updateDocumentStatus)
 );
 
 export { router as applicationsRoutes };
